Date parsing fixes for consistency with Vega.
In particular, Vega mostly uses the JavaScript Date.parse logic to parse dates.
The convention here is that date string that include no mention of time are interpreted
as being at midnight in UTC.  But if time is provided (without an explicit timezone)
then the time is interpreted as local time.